+++ ft-plugin/src/main/java/com/ft/plugin/FTPlugin.java
@@ -14,6 +14,8 @@ import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 
+import java.io.File;
+
 import kotlin.Unit;
 import kotlin.jvm.functions.Function1;
 
@@ -62,11 +64,20 @@ import kotlin.jvm.functions.Function1;
 public class FTPlugin implements Plugin<Project> {
     @Override
     public void apply(Project project) {
-        Logger.init(project);
         FTExtension extension = project.getExtensions().create("FTExt", FTExtension.class, project);
 
         project.afterEvaluate(p -> {
             //Pass parameter object
+            if (extension.enableFileLog) {
+                String logFile;
+                if (extension.logFilePath != null && !extension.logFilePath.isEmpty()) {
+                    logFile = extension.logFilePath;
+                } else {
+                    logFile = p.getBuildDir().getAbsolutePath() + File.separator + "ft-plugin.log";
+                }
+                Logger.init(logFile);
+            }
+
             Logger.setDebug(extension.showLog);
             Logger.debug("Plugin Version:" + BuildConfig.PLUGIN_VERSION +
                     ",ASM Version:" + extension.asmVersion);